+++ spec/selectors/item_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe "microdata" do
+describe "item" do
   before { visit "/microdata.html" }
 
   it "selects microdata" do
@@ -18,5 +18,13 @@ describe "microdata" do
 
   it "selects microdata in a nested types" do
     expect(page).to have_selector :item, "age", type: %w[building person], text: "222"
+  end
+
+  it "does not select an missing type" do
+    expect do
+      expect(page).to have_selector :item, "age", type: "foo", text: "222"
+    end.to raise_error RSpec::Expectations::ExpectationNotMetError, include(<<~EXPECTED.squish)
+      expected to find item "age" within scope foo but there were no matches
+    EXPECTED
   end
 end